feat: config-driven roles, email verification gate, plugin lifecycle hooks, and CSP framing controls
- Add config-driven user roles and decouple registration settings
- Gate login behind mandatory email verification with admin approval
- Fire before_delete and after_unpublish plugin hooks; run before_save on system-fields edits
- Allow iframe embeds via CSP frame-src with configurable same-origin framing
- Keep relative URLs and dedupe imported featured images on import
- Expose post custom fields to theme card templates
- Fix content_aliases FK types for MySQL and user timestamps across databases
- Handle WebP decoding explicitly and surface Hugo import warnings
- Hide Publish button for roles without the publish capability
- Keep generated API keys on one line with a stable copy button
- Modernize Go syntax and expand the homepage features grid

+++ internal/database/migrations/mysql/000004_content_aliases.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS content_aliases (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    content_id BIGINT NOT NULL,
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    content_id INT NOT NULL,
     alias VARCHAR(255) NOT NULL UNIQUE,
     FOREIGN KEY (content_id) REFERENCES content_items(id) ON DELETE CASCADE
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
